[FEAT] Adjust data ratio, Add Impulse test.
accroding to the last trainging, we find that the model behaved poor
predicting with step, impulse input signals.so change the data ratio of
each type to solve it.

after the training finished, start to analyze the model's simulation
of 5th order discrete system. given a unit impulse func (Dirac) to
calculate Frequency Response.